// src/app/interfaces/Playlists.ts
// Generated by https://quicktype.io

export interface Playlists {
    href:     string;
    items:    Item[];
    limit:    number;
    next:     string;
    offset:   number;
    previous: null;
    total:    number;
}

interface Item {
    collaborative: boolean;
    description:   string;
    external_urls: ExternalUrls;
    href:          string;
    id:            string;
    images:        Image[];
    name:          string;
    owner:         Owner;
    primary_color: null;
    public:        boolean;
    snapshot_id:   string;
    tracks:        Tracks;
    type:          string;
    uri:           string;
}

interface ExternalUrls {
    spotify: string;
}

interface Image {
    height: number | null;
    url:    string;
    width:  number | null;
}

interface Owner {
    display_name:  string;
    external_urls: ExternalUrls;
    href:          string;
    id:            string;
    type:          string;
    uri:           string;
}

interface Tracks {
    href:  string;
    total: number;
}